seatwatcher/requirements.docker.txt
alembic==1.13.2
aiosmtplib==3.0.1
click==8.1.7
httpx==0.27.0
jmespath==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.31
tenacity==8.4.1
psycopg2-binary==2.9.9